darwin.apple_sdk.frameworks: add required dependencies
These frameworks tend to themselves include other framework headers; in general
most software expects to see all of them at once, since they're all part of the
macOS SDK.

This commit adds the dependencies required to resolve all of the top-level
framework headers (which appears to be a decent heuristic for those which are
intended to be included). It also introduces a second-layer mechanism by which
the frameworks are checked to ensure that all dependencies are correctly
listed, as well as to bring in dependencies which would not otherwise be
possible due to circular dependencies.